pyproject.toml
[project]
name = "helios-mcp"
version = "0.1.0"
description = "Configuration management for AI behaviors with weighted inheritance"
readme = "README.md"
authors = [
    { name = "Anthony Kougkas", email = "akougkas@iit.edu" }
]
requires-python = ">=3.13"
license = { text = "MIT" }
keywords = ["mcp", "ai", "configuration", "behavior", "inheritance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastmcp>=2.2.6",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/akougkas/helios-mcp"
Repository = "https://github.com/akougkas/helios-mcp"
Issues = "https://github.com/akougkas/helios-mcp/issues"

[project.scripts]
helios-mcp = "helios_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/helios_mcp"]

[tool.ruff]
line-length = 88
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "A", "C4", "SIM", "ARG"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/helios_mcp"]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false